IMGforjingdong/requirements.txt
pillow-avif-plugin==1.4.3
pillow == 10.3.0
requests==2.25.1
pyquery==2.0.0